backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<0.31.0
pydantic>=2.9.2,<3.0.0
python-jose[cryptography]==3.3.0
python-multipart>=0.0.6,<0.0.10
sqlalchemy>=2.0.30,<3.0.0
alembic>=1.13.2,<2.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
fpdf2>=2.7.9,<3.0.0
bcrypt==4.0.1
